Create Makeup & Skin Store Database
In this database information I provide the products I carry in my makeup and skin store. For example, I provide the items, the price, the aisle, the stock amounts we have in the store, and the color of the item. I listed above all of them in correct order according to my database